// backend/src/seeder/enums/permission-permissions.enum.ts
export enum permissionPermissionsEnum {
  CREATE = "create-permission",
  LIST = "list-permission",
  FIND_ONE = "findOne-permission",
  UPDATE = "update-permission",
  DELETE = "delete-permission",
  ADD_ROLE = "addRole-permission",
  REMOVE_ROLE = "removeRole-permission",
  ACCESS_CHECK = "accessCheck-permission",
  HARD_DELETE = "hardDelete-permission",
}